airflow/requirements.txt
faker
pandas
psycopg2-binary
python-dotenv
sqlalchemy